Move the config file code to the config source file
